Data storage 3 and some QoL changes and corrections
Data storage 3 was added a test to show complex data storage in the JSON
file and methods of retrieving that data.

Changed Data Storage 1 test - so it is less spammy in the log. Now the
result only shows on interaction as opposed to continously.

Some updates to the readme files also for sake of minor corrections.